Fixing a bug with public getters. Essentially verisol didn't support public getters for mappings/arrays as it didn't take arguments into account. Have a workaround but only if generating the getters

+++ Sources/SolToBoogie/FunctionEventCollector.cs
@@ -107,7 +107,7 @@ namespace SolToBoogie
                 else if (curType is ArrayTypeName arr)
                 {
                     TypeDescription intDescription = new TypeDescription();
-                    intDescription.TypeString = "uint";
+                    intDescription.TypeString = "uint256";
                     ElementaryTypeName intTypeName = new ElementaryTypeName();
                     intTypeName.TypeDescriptions = intDescription;
                     paramDecl.TypeName = intTypeName;
